backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
python-decouple>=3.8
psycopg2-binary>=2.9.7
django-extensions>=3.2.3
Pillow>=10.0.0
dj-database-url>=2.1.0
gunicorn>=21.2.0
whitenoise>=6.5.0